Remove int-type bug on Windows in skimage.measure.label (#72)
A windows user reported an issue with `skimage.measure.label` in the older `cupyimg` repository: https://github.com/mritools/cupyimg/issues/21.

It seems the same issue is present here, but should be resolved by this PR. I don't think we can easily add a test case on our current linux-based CI as the issue has to do with the NumPy dtype character used to represent the `numpy.int32` dtype. We may want to wait on merging until the fix is confirmed in https://github.com/mritools/cupyimg/issues/21

I think I originally introduced this `int_types` dict with the intention of also supporting >32-bit integers, but it was incomplete and still used `int32` in the actual implementation. This PR just removes this dict altogether and we can revisit the function if we get a request for int64 support.

Authors:
  - Gregory R. Lee (https://github.com/grlee77)

Approvers:
  - https://github.com/jakirkham

URL: https://github.com/rapidsai/cucim/pull/72