Build with NumPy 1.17
Instead of letting `host`'s `numpy` have a minimum, which could float to
a high version that is in turn pinned to a high version in `run`, build
using the minimum required `numpy` version. That way the minimum is set
similarly in `host` and a wide range of `numpy`'s will still work at
runtime.