Apply upstream fix to PiecewiseAffineTransform.estimate
incorporates the change from scikit-image gh-6211

Also updates PiecewiseAffineTransform to be consistent with the other transform types
in keeping self.params as a CuPy array when the inputs are CuPy arrays. Note that
performance of this transform will be very slow due to host/device transfers to
access the CPU-only scipy.spatial.Delaunay.

Co-authored-by: Chris Roat <1053153+chrisroat@users.noreply.github.com>