Promote small integer types to single rather than double precision (#278)
This is a performance-related PR that will result in casting of small integer dtypes (8 and 16-bit) to 32-bit floats rather than the current use of 64-bit floats. The current behavior is consistent with scikit-image, although I have raised an issue there to potentially change to the behavior proposed here: https://github.com/scikit-image/scikit-image/issues/6310. 

The changes required are quite small with the key one being the change to the `new_float_type` dict that gets used when promoting dtypes to a floating point type. Most other changes are in the tests where we sometimes have to bump up the tolerance when computations that were previously in double precision now get run in single precision instead.

I marked this as **breaking** because the output of various floating-point functions for integer inputs may now be `cp.float32` in cases where it was previously `cp.float64`.

Authors:
  - Gregory Lee (https://github.com/grlee77)

Approvers:
  - Gigon Bae (https://github.com/gigony)
  - https://github.com/jakirkham

URL: https://github.com/rapidsai/cucim/pull/278