improve efficiency of histogram-based thresholding functions (#276)
A number of image thresholding functions involve computations on the histogram counts and/or bins. Given that the histogram outputs are small (only 256 elements by default), it is much more efficient to perform computations on them via the CPU.

Overall, computation time of these functions is dominated by the histogram computation itself (at least for larger images), so that is the part it is important to perform on the GPU.

I benchmarked `threshold_otsu` on a (512, 512) grayscale uint8 image (`skimage.data.camera()`). Prior to the change in this PR duration was: 0.894 ms which is reduced to 0.686 ms by the changes made here. The improvement should be around 0.2 ms regardless of image size (since the number of histogram bins is independent of image size).

Authors:
  - Gregory Lee (https://github.com/grlee77)

Approvers:
  - Gigon Bae (https://github.com/gigony)
  - https://github.com/jakirkham

URL: https://github.com/rapidsai/cucim/pull/276