Use in-place multiplication when computing normalized concentrations
We try to avoid in-place modifications of inputs in user-facing functions, but
for this private, internal function it is needed in order to reduce memory
overhead due to making a copy of the potentially large  array.

The change here also avoids unintentional promotion from float32 to float64
that was previously occuring due to multiplication of float32  with
float64 . The in-place multiplication avoids this
potential promotion in precision, preserving memory and improving performance
of the subsequent  call.