Add cucim.skimage.morphology.medial_axis (#342)
closes #336

This PR adds a function for skeletonization of 2D images via the medial axis transform. 

It should be reviewed after #318 is merged. The new commits only start from 19a6fed.

There is one sequential component to this algorithm that still must be run on the CPU, but the majority of the computations are on the GPU and acceleration is good. I will add benchmark results here soon.

Authors:
  - Gregory Lee (https://github.com/grlee77)
  - https://github.com/jakirkham

Approvers:
  - https://github.com/jakirkham
  - Gigon Bae (https://github.com/gigony)

URL: https://github.com/rapidsai/cucim/pull/342