Euclidean distance transform: fix bad choice of block parameters (#393)
closes #392

This PR fixes #392 and also makes it more friendly for use with user-provided `block_params`. In general, most users should not be providing that argument, but it can be used to compare different settings for performance optimization. In case of user-provided `block_params`, the implementation now automatically pad the shape to an appropriate least common multiple of the warp_size and the `m1`, `m2` and `m3` block parameters.

More extensive unit tests over a range of image sizes and `block_params` settings are now implemented.

Authors:
   - Gregory Lee (https://github.com/grlee77)

Approvers:
   - Gigon Bae (https://github.com/gigony)
   - https://github.com/jakirkham