Implement additional deprecations carried out for scikit-image 0.20 (#451)
related to #419

This PR removes most of the deprecated parameters that are being removed in scikit-image 0.20.

There are a few exceptions (e.g. `selem`) where they will be removed in scikit-image 0.20, but we had previously advertised a specific removal date like `2022.03.02`. For those, I will delay the removal until the previously advertised release.

It will be nice to be rid of all the deprecated `multichannel` code! The first, large commit here is just the multichannel deprecation. The rest of the commits are much smaller.

Authors:
  - Gregory Lee (https://github.com/grlee77)

Approvers:
  - Gigon Bae (https://github.com/gigony)

URL: https://github.com/rapidsai/cucim/pull/451