add docs builds to Github Actions (#481)
The PR adds a docs_build process to the PR and Build workflows for this repository. The generated docs are synced to s3 for only the build workflows.

docs build are failing when I tried to run locally. WIP until we figure out the issue. 

edit: docs build was failing due to a newer sphinx version, pinning it to <=5.3.0 fixes the issue

cc @ajschmidt8

Authors:
  - Ajay Thorve (https://github.com/AjayThorve)

Approvers:
  - AJ Schmidt (https://github.com/ajschmidt8)

URL: https://github.com/rapidsai/cucim/pull/481