Binary morphology: omit weights array when possible (#517)
This is a follow-up to #485 allowing the overhead of creating a weights array of uniform values to be omitted when possible for all binary morphology operators. This gives a large performance benefit for small footprints (and small images) with the benefit becoming minimal for large footprint size.

As currently implemented, a shape tuple can be passed to `structure` to indicate the equivalent of an array of ones of that shape. This is a deviation from SciPy's API which does not currently allow this. I am not positive if this is the best API choice or if it might be better to instead add a `size` keyword-only argument like the grayscale morphology functions currently have.

The footprint generation methods `square`, `rectangle`, `cube` etc. use this new tuple format when possible to reduce overhead.

I will post benchmarks below.

Authors:
  - Gregory Lee (https://github.com/grlee77)

Approvers:
  - Gigon Bae (https://github.com/gigony)

URL: https://github.com/rapidsai/cucim/pull/517