Relax `openslide` pin
This was pinned to avoid `openslide` version `4` due to some solver
issues that cropped up with it. Namely it was missing `libdicom` (and an
associated version constraint). However this has been fixed upstream by
adding a repodata patch, adding `run_exports` to `libdicom` (was
missing), and adding a `libdicom` associated test in `openslide`. Given
this we should be able to relax the pin here, which this does.