Add 3rd party license file in Conda packages (#654)
Previously only cuCIM's own license file was included in Conda packages. However the 3rd party license file was not included in Conda packages. This adds that license file as well. Also goes ahead and adds the 3rd party license file to both `libcucim` and `cucim` for good measure (even though it is a little redundant to add it to `cucim` given it depends on `libcucim`, which already has it)

Authors:
  - https://github.com/jakirkham

Approvers:
  - Ray Douglass (https://github.com/raydouglass)
  - Gregory Lee (https://github.com/grlee77)

URL: https://github.com/rapidsai/cucim/pull/654