Require `click` as a wheel dependency
Trying to `import cucim` results in `import click`, which fails if
`click` is not installed. This was encountered recently by a user that
tested out the cuCIM wheel packages.

The Conda packages already solve this by adding `click` to
`requirements/run`. So this makes an equivalent change to the wheel
packages.

Should fix user install issues.