allow path conflicts in conda builds (#801)
Nightly conda builds for 24.12 and 25.02 are currently failing because
of file-clobbering issues with `xorg-*` packages. It looks like this is
because those packages on conda-forge are in the middle of a migration.

That clobbering should generally be safe, and anyway is out of our
control, so this PR proposes temporarily allowing it in CI.

See #800 for details.

## Notes for Reviewers

This is intentionally targeting `branch-24.12`, to get nightly builds
working there. A few more are still needed to complete the 24.12
release.