add batch implementation for various basic regionprops
The functions introduced here are not being added to the public API. They will
be used behind the scenes from `regionprops_table` to enable orders of magnitude
faster computation of region properties for all labels in an image. The basic
approach here is to compute a property for all labels in an image from a single
CUDA kernel call. This is in contrast to the approach from the `RegionProperties`
class which first splits the full image into small sub-images corresponding to
each region and then loops over these small sub-images, computing the requested
property for each small region in turn. That approach is not amenable to good
acceleration on the GPU as individual regions are typically small.

Provides batch implementation that computes the following properties for all properties
in a single kernel call:

- bbox
- label_filled (creates version of label_image with all holes filled)
- num_pixels
- num_pixels_filled
- num_perimeter_pixels (number of pixels at perimeter of each labeled region)
- num_boundary_pixels (number of pixels touching the image boundary for each region)

The following properties are simple transformations of the properties above and
have negligable additional cost to compute:

- area
- area_bbox
- area_filled
- equivalent_diameter_area
- equivalent_spherical_perimeter (as in ITK)
- extent
- perimeter_on_border_ratio (as in ITK)
- slice

The following split the label image into a list of sub-images or subsets of coordinates
where each element in the list corresponds to a label. The background of the label image
has value 0 and is not represented in the sequences. Sequence entry `i` corresponds to
label `i + 1`. In most cases, these will not be needed as properties are now computed
for all regions at once from the labels image, but they are provided for completeness
and to match the scikit-image API.

- coords
- coords_scaled
- image (label mask subimages)
- image_convex (convex label mask subimages)
- image_intensity (intensity_image subimages)
- image_filled (subimages of label mask but with holes filled)
- label (sequence of integer label ids)

Test cases are added that compare the results of these batch computations to results
from scikit-image `regionprops_table`.