Fix path in update-version (#852)
Updates `pyproject.toml` to the correct path.

Technically, the file has nothing that needs to be updated by this
script, but virtually every repo includes updating their
`pyproject.toml` files, so I think it's a good idea to include it anyway
in case the file is changed in the future, there is less chance to
forget to update this script as well.

Skipping CI because this script is not run in CI.