Handle fmt version upgrade
Updated error reporting in CUDA and NVJPEG functions to cast status codes to integers for clearer output. This change enhances the readability of error messages by ensuring that the status codes are displayed as integers rather than their default types.
Without this changes, it causes error, demanding the template method for
the custom types.

Signed-off-by: Gigon Bae <gbae@nvidia.com>