Region Properties Performance Overhaul - Part 6: Public API (regionprops_table) (#848)
Please review #843 first as that explains the general approach in more detail.

## Overview

All of the available region properties were implemented for `regionprops_dict` in parts 1-5. 

This MR's purpose is just to add ability for the existing `cucim.skimage.measure.regionprops_table` function to call `regionprops_dict` by default under the hood. This will allow existing code based on `regionprops_table` to see an immediate performance benefit.

The user can set kwarg `batch_processing=False` to keep using the old implementation based on the `RegionProperties` class. I am not sure if we should provide that option or just remove it. I think the only case it may be useful is that memory requirements could be lower for the non-batch approach as properties are computed on individual regions in isolation. Due to caching behavior of `RegionProperties`, though, the memory usage might not actually be all that much lower.

For compatibility with `regionprops_table` API, the `regionprops_dict` implementation is updated to support user-provided callables for `extra_properties`. For such "extra" properties there will not be a performance benefit to `regionprops_dict` vs. legacy `regionprops_table` as it will also loop over individual regions and apply this function on each.

Authors:
  - Gregory Lee (https://github.com/grlee77)

Approvers:
  - Gigon Bae (https://github.com/gigony)

URL: https://github.com/rapidsai/cucim/pull/848