add a copy of rapids-configure-conda-channels to the repo (#923)
This is the last project in RAPIDS that hasn't converted its conda builds from `conda-build` to `rattler-build`.

That's happening in https://github.com/rapidsai/cucim/pull/864, but that PR isn't yet close to passing.

In https://github.com/rapidsai/gha-tools/issues/145, I'm working on reducing complexity in `gha-tools` by removing unused tools. This repo is the last RAPIDS project using `rapids-configure-conda-channels`, and that usage won't be removed until #864 is complete.

To unblock that cleanup in `gha-tools`, this PR proposes checking in a copy of `rapids-configure-conda-channels` in this repo. That'll allow the work in `gha-tools` to proceed without being blocked by `cucim`'s migration to `rattler-build`.

Authors:
  - James Lamb (https://github.com/jameslamb)

Approvers:
  - Gil Forsyth (https://github.com/gforsyth)

URL: https://github.com/rapidsai/cucim/pull/923