Relax `libcufile` dependency for old CTK 12 on ARM
The CUDA Toolkit (CTK) added cuFile for ARM in CTK 12.2. As a result
always requiring `libcufile` with `libcucim` can make it uninstallable on
ARM with CTK 12.0 & 12.1.

To fix this, build two variants of the `libcucim` package. One with
`libcufile` and one without. Constrain `cuda-version` accordingly so
that `libcufile` is installed whenever possible. Only when `libcufile`
cannot be installed, fallback to the `libcucim` package where the
`libcufile` dependency is left off.

The `x86_64` CTK has had cuFile since the CTK 11.4. As RAPIDS now
requires CUDA 12.0+, there is no issue always requiring `libcufile` on
`x86_64`.

This issue will not be present in CUDA 13+. So always require
`libcufile` in that case.