Include package hash in cuCIM's build strings
To ensure the different cuCIM packages with and without cuFile are both
used, they need different package strings. As the new variant,
`has_cufile`, is included in the package hash, make sure to include the
package hash in the package string. That way each package is named
differently and they don't clobber each other. Thus ensuring both
packages are uploaded and not whichever one was built last.